+++ crates/ibc/src/core/handler.rs
@@ -1,6 +1,6 @@
 use ibc_proto::google::protobuf::Any;
 
-use super::{ics26_routing::msgs::MsgEnvelope, ExecutionContext, RouterError, ValidationContext};
+use super::{msgs::MsgEnvelope, ExecutionContext, RouterError, ValidationContext};
 
 /// Entrypoint which only performs message validation
 ///
@@ -90,7 +90,7 @@ mod tests {
     use crate::core::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
     use crate::core::ics24_host::path::CommitmentPath;
     use crate::core::ics26_routing::context::ModuleId;
-    use crate::core::ics26_routing::msgs::MsgEnvelope;
+    use crate::core::msgs::MsgEnvelope;
     use crate::core::timestamp::Timestamp;
     use crate::core::{dispatch, ValidationContext};
     use crate::mock::client_state::MockClientState;
