+++ crates/ibc/src/core/ics26_routing/mod.rs
@@ -2,4 +2,3 @@
 //! the appropriate module to relay to when a packet is received.
 
 pub mod context;
-pub mod msgs;
