+++ crates/ibc/src/core/mod.rs
@@ -34,6 +34,8 @@ pub mod timestamp;
 
 mod context;
 mod handler;
+// TODO: merge with msgs
+mod msgs;
 mod tx_msg;
 
 pub use handler::dispatch;
@@ -43,6 +45,8 @@ pub use handler::validate;
 pub use context::*;
 
 pub use tx_msg::Msg;
+
+pub use msgs::MsgEnvelope;
 
 pub use ics04_channel::handler::send_packet::{
     send_packet, send_packet_execute, send_packet_validate,
